+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - (Bugfix) Fix Replaced Member Zone during Replace operation
 - (Feature) (ML) Handlers
 - (Feature) Add P0 Compare Func
+- (Bugfix) Handle optional taints for Storage Operator
 
 ## [1.2.35](https://github.com/arangodb/kube-arangodb/tree/1.2.35) (2023-11-06)
 - (Maintenance) Update go-driver to v1.6.0, update IsNotFound() checks